pyproject.toml
[project]
name = "amazon-utils"
version = "0.1.0"
description = "Collection of scripts to use with Amazon Marketplace"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "googletrans>=4.0.2",
    "langdetect>=1.0.9",
    "mistralai>=1.10.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
